pyproject.toml
[tool.poetry]
name = "trello_to_tg_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
py-trello = "^0.18.0"
pyTelegramBotAPI = "^3.8.1"
python-dotenv = "^0.18.0"
marshmallow = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
